Fix missing exception exports and resolve linting issues
Export S3ConcurrencyError and S3ValidationError in __init__.py to resolve
ImportError in phase3_error_handling_demo.py. Also fix line length violation
in examples/common/config.py docstring.
