+++ examples/.env.example
@@ -1,0 +1,31 @@
+# AWS Configuration
+# ================
+
+# Required: S3 bucket name for demo operations
+S3_BUCKET_NAME=your-demo-bucket-name
+
+# Required: AWS region
+AWS_REGION=us-east-1
+
+# Optional: AWS profile (defaults to 'default')
+# AWS_PROFILE=default
+
+# Optional: AWS credentials (if not using profiles or IAM roles)
+# AWS_ACCESS_KEY_ID=your-access-key
+# AWS_SECRET_ACCESS_KEY=your-secret-key
+# AWS_SESSION_TOKEN=your-session-token
+
+# Optional: Custom S3 endpoint (for LocalStack or other S3-compatible services)
+# S3_ENDPOINT_URL=http://localhost:4566
+
+# Feature Flags
+# =============
+
+# Optional: Enable S3 server-side encryption (default: false)
+# S3_ENABLE_ENCRYPTION=true
+
+# Optional: Enable content validation (default: false)
+# S3_ENABLE_VALIDATION=true
+
+# Optional: Enable request logging (default: false)
+# S3_ENABLE_LOGGING=true
