--- examples/s3-artifact-demo/.env.example
@@ -1,13 +1,0 @@
-# AWS Credentials
-AWS_ACCESS_KEY_ID=your_access_key_here
-AWS_SECRET_ACCESS_KEY=your_secret_key_here
-AWS_REGION=us-east-1
-
-# S3 Configuration
-S3_BUCKET_NAME=your-artifact-bucket
-
-# Optional: For S3-compatible services (MinIO, LocalStack, etc.)
-# S3_ENDPOINT_URL=https://your-s3-compatible-service.com
-
-# Optional: AWS Session Token for temporary credentials
-# AWS_SESSION_TOKEN=your_session_token_here
