--- examples/s3-artifact-demo/pyproject.toml
@@ -1,40 +1,0 @@
-[project]
-name = "s3-artifact-demo"
-version = "0.1.0"
-description = "Demonstration of AWS S3 artifact service integration with Google ADK"
-authors = [{ name = "Google Cloud", email = "noreply@google.com" }]
-license = "Apache-2.0"
-readme = "README.md"
-requires-python = ">=3.9"
-dependencies = [
-    "google-adk>=1.0.0",
-    "boto3>=1.34.0",
-    "python-dotenv>=1.0.0",
-    "pydantic>=2.0.0",
-]
-
-[project.optional-dependencies]
-dev = [
-    "pytest>=8.0.0",
-    "pytest-asyncio>=0.25.0",
-    "pytest-cov>=6.0.0",
-    "pyink>=24.0.0",
-    "pylint>=3.0.0",
-]
-
-[build-system]
-requires = ["setuptools>=45", "wheel"]
-build-backend = "setuptools.build_meta"
-
-[tool.pytest.ini_options]
-console_output_style = "progress"
-addopts = "-vv -s"
-testpaths = ["tests/"]
-markers = ["unit", "integration"]
-log_level = "ERROR"
-filterwarnings = ["ignore::UserWarning"]
-
-[tool.pyink]
-line-length = 80
-pyink-indentation = 4
-pyink-use-majority-quotes = true
