+++ pyproject.toml
@@ -27,6 +27,7 @@ dependencies = [
     "boto3>=1.34.0",
     "botocore>=1.34.0",
     "cryptography>=41.0.0",
+    "python-dotenv>=1.0.0",
 ]
 
 [project.optional-dependencies]
@@ -94,7 +95,7 @@ module = ["mcp.*", "google.*", "boto3.*", "botocore.*"]
 ignore_missing_imports = true
 
 [tool.pytest.ini_options]
-testpaths = ["tests", "examples/s3-artifact-demo/tests"]
+testpaths = ["tests"]
 python_files = ["test_*.py"]
 python_classes = ["Test*"]
 python_functions = ["test_*"]
